# be/api/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  fiber.Map:
    additionalProperties: true
    type: object
...
  http.RateMaterialRequest:
    properties:
      review:
        maxLength: 2000
        type: string
      score:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - score
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@ngasihtau.com
    name: NgasihTau API Support
  description: |-
    NgasihTau is a learning platform API that enables teachers to create Knowledge Pods for sharing learning materials with students.
    The platform uses a microservices architecture with Go as the primary backend language.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: NgasihTau API
  version: "1.0"
paths:
...
  /materials/{id}/ratings:
    get:
      consumes:
      - application/json
      description: Get ratings for a material. Use summary=true for rating summary.
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Return summary instead of list
        in: query
        name: summary
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        maximum: 50
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ratings or summary
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid material ID
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Get ratings
      tags:
      - Ratings
    post:
      consumes:
      - application/json
      description: Rate a material (1-5 stars) with optional review. Updates existing
        rating if already rated.
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Rating and optional review
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.RateMaterialRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created/updated rating
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Rate a material
      tags:
      - Ratings
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"